Phillips feat 10 (#11)
* feat: LOGIN SCRATCH (80%)

* feat: LOGIN DONE

* feat/TKAPAP-10-SIFACTORY

* feature 10 done

+++ SI-FACTORY/src/main/java/apap/group/assignment/SIFACTORY/repository/ProduksiDB.java
@@ -4,5 +4,5 @@ import apap.group.assignment.SIFACTORY.model.ProduksiModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProduksiDB extends JpaRepository<ProduksiModel, Long> {
-    ProduksiModel findByIdProduksi(Integer id);
+    ProduksiModel findByIdProduksi(Long id);
 }
